// src/generated/prisma/enums.ts
export const ActivityEntityType = {
  CLIENT: 'CLIENT',
  PROJECT: 'PROJECT',
  TASK: 'TASK',
  INVOICE: 'INVOICE',
  MEMBERSHIP: 'MEMBERSHIP',
  INVITATION: 'INVITATION',
  ATTACHMENT: 'ATTACHMENT',
  PORTAL_USER: 'PORTAL_USER',
  COMMENT: 'COMMENT'
} as const

export type ActivityEntityType = (typeof ActivityEntityType)[keyof typeof ActivityEntityType]


export const ActivityAction = {
  CREATED: 'CREATED',
  UPDATED: 'UPDATED',
  STATUS_CHANGED: 'STATUS_CHANGED',
  DELETED: 'DELETED',
  INVITATION_SENT: 'INVITATION_SENT',
  INVITATION_RESENT: 'INVITATION_RESENT',
  INVITATION_CANCELED: 'INVITATION_CANCELED',
  INVITATION_ACCEPTED: 'INVITATION_ACCEPTED',
  ROLE_CHANGED: 'ROLE_CHANGED',
  OWNERSHIP_TRANSFERRED: 'OWNERSHIP_TRANSFERRED',
  MEMBER_REMOVED: 'MEMBER_REMOVED',
  MEMBER_LEFT: 'MEMBER_LEFT',
  FILE_UPLOADED: 'FILE_UPLOADED',
  FILE_DELETED: 'FILE_DELETED',
  PORTAL_INVITATION_SENT: 'PORTAL_INVITATION_SENT',
  PORTAL_INVITATION_RESENT: 'PORTAL_INVITATION_RESENT',
  PORTAL_INVITATION_CANCELED: 'PORTAL_INVITATION_CANCELED',
  PORTAL_INVITATION_ACCEPTED: 'PORTAL_INVITATION_ACCEPTED',
  PORTAL_USER_REMOVED: 'PORTAL_USER_REMOVED'
} as const

export type ActivityAction = (typeof ActivityAction)[keyof typeof ActivityAction]